+++ HISTORY.rst
@@ -3,6 +3,13 @@
 Release History
 ---------------
 
+Next Release
+++++++++
+
+**New Features and Enhancements:**
+
+- Add metadata query functionality (`#574 <https://github.com/box/box-python-sdk/pull/574>`_)
+
 2.11.0 (2021-01-11)
 ++++++++